+++ scripts/build.sh
@@ -5,20 +5,22 @@ export RUST_LOG=info
 # for printing configuration
 python -m kgdata.config
 export LOG_CONFIG=0  # surrpress logging directory configurations
+# export BUILD_DATASET_ARGS="--verify-signature"
+export BUILD_DATASET_ARGS=""
 
 function dbpedia_dataset {
     echo "Build dbpedia dataset: $1"
-    python -m kgdata.dbpedia.datasets -d $1
+    python -m kgdata.dbpedia.datasets $BUILD_DATASET_ARGS -d $1
 }
 
 function wikidata_dataset {
     echo "Build wikidata dataset: $1"
-    python -m kgdata.wikidata.datasets -d $1 ${@:2}
+    python -m kgdata.wikidata.datasets $BUILD_DATASET_ARGS -d $1 ${@:2}
 }
 
 function wikipedia_dataset {
     echo "Build wikipedia dataset: $1"
-    python -m kgdata.wikipedia.datasets -d $1 ${@:2}
+    python -m kgdata.wikipedia.datasets $BUILD_DATASET_ARGS -d $1 ${@:2}
 }
 
 function dbpedia_db {
@@ -38,7 +40,7 @@ function wikidata_db {
 # dbpedia_dataset mapping_extractor_dump
 # dbpedia_dataset ontology_dump
 # dbpedia_dataset classes
-# dbpedia_dataset properties
+dbpedia_dataset properties
 # dbpedia_dataset entities
 # dbpedia_dataset entity_redirections
 # dbpedia_dataset entity_labels
@@ -58,6 +60,7 @@ function wikidata_db {
 # # python -m kgdata.wikidata.datasets -d entity_redirection_dump --sign
 # # python -m kgdata.wikidata.datasets -d page_dump --sign
 
+# ----- verified dataset -----
 # wikidata_dataset page_ids
 # wikidata_dataset entity_ids
 # wikidata_dataset entity_redirections
@@ -65,14 +68,12 @@ function wikidata_db {
 # wikidata_dataset entity_types
 
 # wikidata_dataset classes
-wikidata_dataset properties
+# wikidata_dataset properties
 
 # wikidata_dataset class_count
 # wikidata_dataset property_count
 # wikidata_dataset property_domains
 # wikidata_dataset property_ranges
-
-# wikidata_dataset cross_wiki_mapping
 
 # wikidata_dataset entity_metadata
 # wikidata_dataset entity_all_types
@@ -83,8 +84,11 @@ wikidata_dataset properties
 # wikidata_dataset entity_pagerank
 # wikidata_dataset entity_wiki_aliases
 
+# wikidata_dataset cross_wiki_mapping
 # wikidata_dataset meta_graph
 # wikidata_dataset meta_graph_stats
+
+# ----- unverified dataset -----
 
 # ======================================================================
 # WIKIPEDIA Datasets
@@ -95,6 +99,7 @@ wikidata_dataset properties
 # wikipedia_dataset article_links
 # wikipedia_dataset article_metadata
 # wikipedia_dataset html_tables
+
 # wikipedia_dataset relational_tables
 # wikipedia_dataset linked_relational_tables
 # wikipedia_dataset easy_tables
@@ -104,7 +109,7 @@ wikidata_dataset properties
 # DBpedia Databases
 
 # dbpedia_db classes
-# dbpedia_db properties
+dbpedia_db properties
 # dbpedia_db entities
 # dbpedia_db entity_labels
 # dbpedia_db entity_metadata
@@ -114,7 +119,7 @@ wikidata_dataset properties
 # WIKIDATA Databases
 
 # wikidata_db classes
-wikidata_db properties
+# wikidata_db properties
 # wikidata_db entities
 # wikidata_db entity_labels
 # wikidata_db entity_metadata